\name{ausplots_trajectory}
\alias{ausplots_trajectory}
\title{Visualisation of species composition and community index revisit trajectories
}
\description{Generates 2D plots visualising the trajectory of species composition (PCoA) and diversity (scatterplot) for revisit series. Trajectories are calculated from the TERN AusPlots point intercept module (\code{$veg.PI}) as generated by \code{\link{get_ausplots}} and other pre-processing functions.
}
\usage{
ausplots_trajectory(my.ausplots.object, choices = c("PCoA", "diversity"), min.revisits = 3, plot_select = NULL)
}
\arguments{
  \item{my.ausplots.object}{Input raw data object (list) as generated in a call from the \code{\link{get_ausplots}} function.
}
  \item{choices}{Character. User selection of PCoA and/or diversity plots. Both are generated by default.
}
  \item{min.revisits}{Integer. Minimum number of revisits required to include a plot. 
}
  \item{plot_select}{Optional character vector of AusPlots plot IDs (\code{site_location_name}) to request data for a specific set of plots.
}
}
\details{This function applies the functionality of R package \code{ecotraj} (De C치ceres et al. 2019; Sturbois et al. 2021) to the TERN AusPlots vegetation point intercept data by including the required pre-processing of the raw data and modifications to the output plot format. The function is intended as a rapid, visual means to compare revisit trajectories in species composition and diversity among sites. For in-depth change/trajectory analysis, users are directed to work with the raw data and the extensive functionality in \code{ecotraj} or other approaches.

The default graphical outputs require \code{$site.info}, to order visits by date, and \code{$veg.PI}, for plant species abundance, in the AusPlots data extract.

Two output options are available. 'PCoA' displays the first two axes (and their respective deviance explained) of a metric ordination based on species abundances and so shows the trajectory of plots in species composition space. Similar trajectories on the plot implies parallel trends in species presence and abundance. Next, 'diversity' displays a simple scatterplot with trajectories between revisits. The function presently computes species richness and Shannon diversity metrics on the 2D plot. With pending updates to the data tables available via \code{\link{get_ausplots}}, users will also be able to plot out the trajectory of other community level indices such as community temperature index, proportional abundance of C4 photosynthetic pathway, species abundance distribution (sigma parameter of the lognormal distribution) and functional trait moments, for example.

Note that revisits (or revisit series > 2) are available only for a subset of the full plot network and that \code{ausplots_trajectory} may display trajectories better for targeted sets of plots.
}
\value{
The output is plotted to the plot device.}
\references{
De C치ceres, M., Coll, L., Legendre, P., Allen, R.B., Wiser, S.K., Fortin, M.J., Condit, R. & Hubbell, S. (2019). Trajectory analysis in community ecology. Ecological Monographs 89: e01350

Sturbois, A., De C치ceres, M., S치nchez-Pinillos, M., Schaal, G., Gauthier, O., Le Mao, P., Ponsero, A., & Desroy, N. (2021). Extending community trajectory analysis : New metrics and representation. Ecological Modelling 440: 109400
}
\author{Greg Guerin
}
\seealso{\code{\link{get_ausplots}}
  \cr
  \cr
	\code{\link{ausplots_visual}} 
}
\examples{
\dontrun{
#Extract site data first:
my.data <- get_ausplots()

#Extract vegetation point intercept data for plots with revisits
#Here, subsetting my.Plots_IDs to revisits in one bioregion
my.data <- get_ausplots(veg.PI = TRUE, 
my.Plot_IDs = unique(subset(my.data$site.info, visit_number > 1 
& bioregion_name == "MAC", select=site_location_name))[,1])

#Display species composition trajectory for all plots with at least 3 revisits
ausplots_trajectory(my.data, choices = "PCoA") 
}
}
