\name{species_table}
\alias{species_table}
\title{Species occurrence matrices for AusPlots
}
\description{This function takes a data frame of individual raw point-intercept hits from AusPlots generated using the \code{\link{get_ausplots}} function, and generates species occurrence matrices as desired based on presence/absence, cover, frequencey or IVI.
}
\usage{
species_table(veg.PI, m_kind=c("PA", "percent_cover", "freq", "IVI"),
	cover_type=c("PFC", "OCC"))
}
\arguments{
  \item{veg.PI}{The input raw point-intercept data generated by \code{\link{get_ausplots}}
}
  \item{m_kind}{The desired species scoring method: PA: binary (presence/absence), percent cover, frequency (based on occurrences on the 10 point-intercept transects within plots) or IVI (combination of cover and frequency - see Details).
}
  \item{cover_type}{A choice between PFC: 'projected foliage cover' and OCC: 'opaque canopy cover' and only applies to percent_cover and IVI. If 'PFC' is selected, hits scored as 'in canopy sky' are removed, whereas they are retained as cover for that species for 'OCC' (which may be more suitable for some applications such as comparison to remotely sensed cover products).
 }
}
\details{Species occurrence matrices are the basis for a wide range of downstream analyses in community ecology, such as classification, ordination, alpha and beta diversity, abundance distributions and many more. Many of these analyses can be performed in the vegan R package using the returned matrix.

IVI (Importance Value Index; Ellenberg & Mueller-Dombois 1974) is calculated as the sum of relative abundance (percent cover) and relative frequency (frequency of occurrence on the 10 transects used to gather the point-intercept data). It has a theoretical range of 0 - 200 and is calculated as follows:

\deqn{(%cover / total %cover)*100 + (%frequency / total %frequency)*100}

The returned occurrence matrix (species against sites) can be converted to 'long' format (with rows as species records) using:

\code{my.matrix <- species_table(my.veg.PI, m_kind="PA", cover_type="PFC")}
\cr
\code{library(reshape2)}
\cr
\code{my.matrix <- as.data.frame(melt(as.matrix(my.matrix)))}
\cr
\code{my.matrix <- my.matrix[-which(my.matrix $value == 0),]}
\cr
\cr

The \code{\link{round}} function can be used on the returned matrix to estimate percent covers to the nearest percent if desired (i.e., with digits=0)
}

\value{Returns a matrix with species as columns and sites as rows as a data frame. Values are determined by argument \code{m_kind}.
}

\author{Greg R. Guerin
}

\references{
Ellenberg, D. & Mueller-Dombois, D. (1974) Aims and methods of vegetation ecology. Wiley, New York, NY.
}

\seealso{
\code{\link{get_ausplots}}
}
\examples{
#get veg.PI data for selected plots:
my.data <- get_ausplots(my.Plot_IDs=c("SATFLB0004", "QDAMGD0022", "NTASTU0002"),
	veg.vouchers=FALSE, site_info=FALSE)

#generate a percent cover occurrence matrix:
species_table(my.data$veg.PI, m_kind="percent_cover", cover_type="PFC")
}

