\name{growth_form_table}
\alias{growth_form_table}
\title{Plant growth form occurrence matrices for TERN AusPlots
}
\description{Generates occurrence matrices for NVIS plant growth forms in plots as desired based on presence/absence, percent cover or species richness (i.e., the number of species assigned to a particular growth form). The input is a data frame of raw point intercept data from AusPlots generated using the \code{get_ausplots} function.
}
\usage{
growth_form_table(veg.PI, m_kind=c("PA", "percent_cover", "richness"),
	cover_type=c("PFC", "OCC"))
}
\arguments{
  \item{veg.PI}{The input raw point intercept data generated by \code{\link{get_ausplots}}
}
  \item{m_kind}{The desired scoring method: binary (presence/absence; PA), percent cover or species richness.
}
  \item{cover_type}{A choice between PFC: 'projected foliage cover' and OCC: 'opaque canopy cover' and only applies to percent_cover. If 'PFC' is selected, hits scored as 'in canopy sky' are removed, whereas they are retained as cover for that species for 'OCC'
}
}
\details{Probably the most useful data output from this function is the matrix of plant growth forms against sites with percent cover. This can be used in manova, ordination, classification etc at continental scale where species turnover is too high for some methods to provide meaningful results. It also provides a multivariate structural and functional response variable.
	
	Note sometimes the same species is recorded with different growth forms in a plot and therefore the rowSums of the occurrence matrix when m_kind is set to "richness" can be higher than the observed species richness because the same species can count towards the weights for multiple growth forms.
	
	The returned occurrence matrix (plant growth forms against sites) can be converted to long format (with rows as individual occurrences of growth forms) using:

\code{

my.matrix <- growth_form_table(my.veg.PI, m_kind="percent_cover", cover_type="PFC"); 
library(reshape2); 
my.matrix <- as.data.frame(melt(as.matrix(my.matrix))); 
my.matrix <- my.matrix[-which(my.matrix $value == 0),]

}

The \code{\link{round}} function can be used on the returned matrix to estimate percent covers to the nearest percent if desired (i.e., with digits=0)

}
\value{Returns a matrix with plant growth forms as columns and sites as rows as a data frame. Values are determined by argument m_kind
}
\author{Greg R. Guerin
}

\seealso{\code{\link{get_ausplots}}
}
\examples{
#get veg.PI data for selected plots:
my.data <- get_ausplots(my.Plot_IDs=c("SATFLB0004", "QDAMGD0022", "NTASTU0002"),
	veg.vouchers=FALSE, site_info=FALSE)

#generate a percent cover occurrence matrix for plant growth forms:
growth_form_table(my.data$veg.PI, m_kind="percent_cover", cover_type="PFC")
}
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
