\name{species_table}
\alias{species_table}
\title{Species occurrence matrices for AusPlots
}
\description{This function takes a data frame of individual raw point intercept hits from AusPlots generated using the \code{get_ausplots} function, and generates species occurrence matrices as desired based on presence/absence, cover, frequencey or IVI index.
}
\usage{
species_table(veg.PI, m_kind=c("PA", "percent_cover", "freq", "IVI"),
	cover_type=c("PFC", "OCC"), species_name=c("HD","SN","GS"))
}
\arguments{
  \item{veg.PI}{The input raw point intercept data generated by \code{\link{get_ausplots}}
}
  \item{m_kind}{The desired species scoring method: PA: binary (presence/absence), percent cover, frequency (based on occurrences on transects within plots) or IVI (combination of cover and frequency).
}
  \item{cover_type}{A choice between PFC: 'projected foliage cover' and OCC: 'opaque canopy cover' and only applies to percent_cover and IVI. If 'PFC' is selected, hits scored as 'in canopy sky' are removed, whereas they are retained as cover for that species for 'OCC' (which may be more suitable for some applications such as comparison to remotely sensed cover products).
 }

\item{species_name}{Set to "SN" (default) to use "standardised_names"", "HD" to use "herbarium_determination"", and "GS" to use "genus_species". See details for explaination of each setting.
 
}
\details{\code{species_name} provides multiple options for calculating plant cover. If \code{species_name=c("HD")}, the function will use the "herbarium_determination" provided by state herbaria to indentify unique species. Herbarium determinations are the most commonly used scientific names in the given state and vary between states due to disagreements on taxonomy/nomenclature. Herbarium determinations are also made to the lowest possible taxonomic rank, which may be the family, genus, species, subspecies or variety. These attributes may interfere with between-plot comparisons. If \code{species_name=c("SN")}, the default, scientific names are based on herbarium determinations but have been edited to match the most commonly used or widely accepted scientific name in Australia, referred to as the "standardised_name"" in the veg.PI input table. This helps ensure scientific names are consistent between plots in different states. However, standardised_names are also assessed to the lowest possible taxonomic rank. If \code{species_name=c("GS")}, only the standardised Genus species rank ("genus_species" in the veg.PI input table) is used. This helps maximise comparability and consistency between plots, but reduces the taxonomic detail.

The returned occurrence matrix (species against sites) can be converted to long format (with rows as species records) using:

\code{my.matrix <- species_table(my.veg.PI, m_kind="PA", cover_type="PFC")}
\cr
\code{library(reshape2)}
\cr
\code{my.matrix <- as.data.frame(melt(as.matrix(my.matrix)))}
\cr
\code{my.matrix <- my.matrix[-which(my.matrix $value == 0),]}
\cr
\cr

The \code{\link{round}} function can be used on the returned matrix to estimate percent covers to the nearest percent if desired (i.e., with digits=0)
}

\value{Returns a matrix with species as columns and sites as rows as a data frame. Values are determined by argument m_kind
}

\author{Greg R. Guerin
}


\seealso{
\code{\link{get_ausplots}}
}
\examples{
#get veg.PI data for selected plots:
my.data <- get_ausplots(my.Plot_IDs=c("SATFLB0004", "QDAMGD0022", "NTASTU0002"),
	veg.vouchers=FALSE, site_info=FALSE)

#generate a percent cover occurrence matrix:
species_table(my.data$veg.PI, m_kind="percent_cover", cover_type="PFC")
}

